moved string creation helper to utils
